Create fix_taxonomy.py
It was created to fix some missing classification provided by Silva database.
For better understanding:

Consider one group with 5 persons:
Lucy Lee; Lucy Bee; Lucy (unknown); Bob Carter; Jeff Bridge

Names as higher taxa groups and Forenames lower taxa. But to classify all Lucy I am obligated to have two words (name and forename). If we have "Lucy" as one higher group, independent of their forenames, they are "Lucy's members". So, if one Lucy's their forename as unknown, but still Lucy, she will belong to the same group as the others. to fix this in other to include Lucy (unknown) as a member and do not discharge her, I supplemented her name with 'forename'. In this way, Lucy has a complete name and forename to be classified.